Added diesel orm to api

+++ api/src/routes.rs
@@ -2,9 +2,10 @@ use crate::handlers;
 use crate::APP_URL;
 use warp::Filter;
 
-// A function to build our routes
 pub fn routes() -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
     get_post()
+        // Add other routes here
+        // .or(other_route())
         .with(
             warp::cors()
                 .allow_origin(APP_URL)
@@ -15,7 +16,7 @@ pub fn routes() -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejecti
 
 // A route to handle GET requests for a specific post
 fn get_post() -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-    warp::path!("posts" / u64)
+    warp::path!("posts" / i32)
         .and(warp::get())
         .and_then(handlers::get_post)
 }
